+++ packages/frontend/src/downstream/SearchInput.tsx
@@ -2,14 +2,8 @@ import styled from "styled-components";
 import { Input } from "semantic-ui-react";
 
 const SearchInput = styled(Input)`
-  ${props =>
-    props.theme.pristine
-      ? `
-    margin-bottom: 1em;`
-      : `
-  margin-top: 3em;
-  margin-bottom: 3em;
-  `}
+  width: 100%;
+  margin-right: 1em;
 `;
 
 export { SearchInput };