+++ packages/frontend/src/index.tsx
@@ -1,5 +1,11 @@
 import * as React from "react";
 import { render } from "react-dom";
 import { RoutesRoute } from "./routes/Routes";
+import { Provider } from "unstated";
 
-render(<RoutesRoute />, document.getElementById("root"));
+render(
+  <Provider>
+    <RoutesRoute />
+  </Provider>,
+  document.getElementById("root")
+);