+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * feat(core, services, gateway): Add context action to english service ([f4e275b](https://gitlab.com/pojntfx/learn-chinese-platform/commit/f4e275b))
 * feat(core, services): Add getContextForHanzi action to context service ([463e83a](https://gitlab.com/pojntfx/learn-chinese-platform/commit/463e83a))
+* feat(core, services): Add universal context ([fc39bf1](https://gitlab.com/pojntfx/learn-chinese-platform/commit/fc39bf1))