+++ CHANGELOG.md
@@ -1,9 +1,15 @@
-## <small>0.0.1-0-monorepo (2018-12-11)</small>
+## <small>0.0.1-0-monorepo (2018-12-12)</small>
 
-* feat(frontend): Add precision toggle ([24efb78](https://gitlab.com/pojntfx/learn-chinese-platform/commit/24efb78))
+
+
+
+## <small>0.0.1-alpha.15 (2018-12-11)</small>
+
 * feat(core, services, gateway): Add context action to english service ([f4e275b](https://gitlab.com/pojntfx/learn-chinese-platform/commit/f4e275b))
 * feat(core, services): Add getContextForHanzi action to context service ([463e83a](https://gitlab.com/pojntfx/learn-chinese-platform/commit/463e83a))
 * feat(core, services): Add universal context ([fc39bf1](https://gitlab.com/pojntfx/learn-chinese-platform/commit/fc39bf1))
+* v0.0.1-alpha.15 ([60ab6b9](https://gitlab.com/pojntfx/learn-chinese-platform/commit/60ab6b9))
+* feat(frontend): Add precision toggle ([24efb78](https://gitlab.com/pojntfx/learn-chinese-platform/commit/24efb78))